Changed realization - img_ & res_ vectors now store uint8_t type - removed check for bounds [0,255] in ValidationImpl() as unnecessary

+++ tasks/seq/varfolomeev_g_histogram_linear_stretching/include/ops_seq.hpp
@@ -16,7 +16,7 @@ class TestTaskSequential : public ppc::core::Task {
   bool PostProcessingImpl() override;
 
  private:
-  std::vector<int> img_, res_;
+  std::vector<uint8_t> img_, res_;
 };
 
 }  // namespace varfolomeev_g_histogram_linear_stretching_seq
